pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "haotian"
version = "0.1.0"
description = "Haotian skill-first capability intelligence pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "beautifulsoup4>=4.12,<5.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.6,<3.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0,<9.0",
]

[project.scripts]
haotian = "haotian.main:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
